+++ ios/IssueTracker/IssueTracker/Network/LoginManager.swift
@@ -10,29 +10,7 @@ import Combine
 
 class LoginManager {
     
-    private let callbackUrlScheme = "issueTracker"
     
-    enum LoginURL {
-        static let scheme = "https"
-        static let host = "github.com"
-        static let path = "/login/oauth/authorize"
-        static let clientID = "d" /// 민감한 정보 숨기기
-        static let redirectURI = "issueTracker://login"
-        static let scope = "user"
-        
-        static func url() -> URL? {
-            var components = URLComponents()
-            components.scheme = LoginURL.scheme
-            components.host = LoginURL.host
-            components.path = LoginURL.path
-            components.queryItems = [
-                URLQueryItem(name: "client_id", value: clientID),
-                URLQueryItem(name: "redirect_uri", value: redirectURI),
-                URLQueryItem(name: "scope", value: scope)
-            ]
-            return components.url
-        }
-    }
 }
 
 
@@ -40,7 +18,7 @@ extension LoginManager {
     
     func requestCode(handler: @escaping (URL, String)->()) {
         let url = LoginURL.url()!
-        handler(url, callbackUrlScheme)
+        handler(url, LoginURL.callbackUrlScheme)
     }
     
 }
